requirements.txt
python-dotenv==0.19.2
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0 
psycopg2-binary==2.8.6